downloader_extra/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
fastapi==0.135.2
greenlet==3.3.2
h11==0.16.0
idna==3.11
numpy==2.4.3
pandas==3.0.1
polars==1.39.3
polars-runtime-32==1.39.3
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyyaml==6.0.3
requests==2.33.0
six==1.17.0
sqlalchemy==2.0.48
starlette==1.0.0
tabulate==0.10.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
wbgapi==1.0.14
